- add course backend and UI - migration and seeder for courses, course_details, course_images, course_options, hospital_images, options - bug fix

+++ database/migrations/2019_05_02_094857_create_courses_table.php
@@ -37,7 +37,7 @@ class CreateCoursesTable extends Migration
             $table->string('price_memo',50)->nullable();
             $table->integer('regular_price')->nullable();
             $table->integer('discounted_p[rice')->nullable();
-            $table->tinyInteger('tax_class')->unsigned()->default(0);
+            $table->integer('tax_class')->unsigned()->default(0);
             $table->tinyInteger('display_setting')->nullable();
             $table->integer('pv')->unsigned()->default(0);
             $table->integer('pvad')->unsigned()->default(0);
